pyproject.toml
[project]
name = "astro-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
